pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["gm*"]
exclude = ["tests*"]

[project]
name = "gm"
version = "0.1.0"
description = "Git Worktree Manager - Professional Git Worktree Management Tool"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "GM Team", email = "dev@gm-tool.io"},
]
license = "MIT"

dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "structlog>=22.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.6.0",
    "pytest-cov>=3.0.0",
    "mypy>=0.950",
    "black>=22.0.0",
    "ruff>=0.0.200",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "sphinx-click>=4.0.0",
]

[project.scripts]
gm = "gm.cli.main:cli"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "cli: marks tests as CLI tests",
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "W"]
ignore = ["E203", "E266", "E501", "W503"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
